+++ .rebase/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 The file to keep a list of changed files which will potentionaly help to resolve rebase conflicts.
 
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/559
+
+- code/src/vs/workbench/contrib/webview/browser/pre/index.html
+- code/src/vs/platform/extensionManagement/common/extensionManagement.ts
+- code/src/vs/platform/extensionManagement/common/extensionGalleryService.ts
+- 
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/572
+
+- code/package.json
+- code/build/package.json
+- code/remote/package.json
+---
+
 #### @sbouchet
 https://github.com/che-incubator/che-code/pull/571
 
@@ -72,7 +89,6 @@ https://github.com/che-incubator/che-code/pull/528
 https://github.com/che-incubator/che-code/pull/526
 
 - code/package.json
-- code/src/vs/platform/utilityProcess/electron-main/utilityProcess.ts
 ---
 
 #### @svor
@@ -212,7 +228,6 @@ https://github.com/che-incubator/che-code/commit/eed0a5213ba1b29b810d53f6365aaa2
 #### @vitaliy-guliy
 https://github.com/che-incubator/che-code/pull/211
 
-- code/src/vs/workbench/contrib/webview/browser/pre/index-no-csp.html
 - code/src/vs/workbench/contrib/webview/browser/pre/index.html
 ---